+++ mRemoteNG/Connection/Protocol/PuttyBase.cs
@@ -11,6 +11,7 @@ using System.Threading;
 using System.Windows.Forms;
 using mRemoteNG.Properties;
 using mRemoteNG.Resources.Language;
+using SecretServerInterface;
 
 // ReSharper disable ArrangeAccessorOwnerBody
 
@@ -111,6 +112,20 @@ namespace mRemoteNG.Connection.Protocol
                                 var cryptographyProvider = new LegacyRijndaelCryptographyProvider();
                                 password = cryptographyProvider.Decrypt(Settings.Default.DefaultPassword,
                                                                         Runtime.EncryptionKey);
+                            }
+                        }
+
+                        // access secret server api if necessary
+                        if (username.StartsWith("SSAPI:"))
+                        {
+                            var domain = ""; // dummy
+                            try
+                            {
+                                SecretServerInterface.SecretServerInterface.fetchSecretFromServer(username, out username, out password, out domain);
+                            }
+                            catch (Exception ex)
+                            {
+                                Event_ErrorOccured(this, "Secret Server Interface Error: " + ex.Message, 0);
                             }
                         }